Mirroring the Alluxio worker class hierarchy for the master.
