+++ core/server/src/main/java/alluxio/master/block/BlockMasterClientRestServiceHandler.java
@@ -13,7 +13,7 @@ package alluxio.master.block;
 
 import alluxio.Constants;
 import alluxio.RestUtils;
-import alluxio.master.AlluxioMaster;
+import alluxio.master.AlluxioMasterService;
 import alluxio.web.MasterWebServer;
 import alluxio.wire.BlockInfo;
 
@@ -54,9 +54,8 @@ public final class BlockMasterClientRestServiceHandler {
    */
   public BlockMasterClientRestServiceHandler(@Context ServletContext context) {
     // Poor man's dependency injection through the Jersey application scope.
-    AlluxioMaster master =
-        (AlluxioMaster) context.getAttribute(MasterWebServer.ALLUXIO_MASTER_SERVLET_RESOURCE_KEY);
-    mBlockMaster = master.getBlockMaster();
+    mBlockMaster = ((AlluxioMasterService) context
+        .getAttribute(MasterWebServer.ALLUXIO_MASTER_SERVLET_RESOURCE_KEY)).getBlockMaster();
   }
 
   /**
