+++ core/server/src/main/java/alluxio/master/lineage/LineageMasterClientRestServiceHandler.java
@@ -16,7 +16,7 @@ import alluxio.Constants;
 import alluxio.RestUtils;
 import alluxio.job.CommandLineJob;
 import alluxio.job.JobConf;
-import alluxio.master.AlluxioMaster;
+import alluxio.master.AlluxioMasterService;
 import alluxio.web.MasterWebServer;
 import alluxio.wire.LineageInfo;
 import alluxio.wire.TtlAction;
@@ -66,9 +66,8 @@ public final class LineageMasterClientRestServiceHandler {
    */
   public LineageMasterClientRestServiceHandler(@Context ServletContext context) {
     // Poor man's dependency injection through the Jersey application scope.
-    AlluxioMaster master =
-        (AlluxioMaster) context.getAttribute(MasterWebServer.ALLUXIO_MASTER_SERVLET_RESOURCE_KEY);
-    mLineageMaster = master.getLineageMaster();
+    mLineageMaster = ((AlluxioMasterService) context
+        .getAttribute(MasterWebServer.ALLUXIO_MASTER_SERVLET_RESOURCE_KEY)).getLineageMaster();
   }
 
   /**
