+++ core/server/src/main/java/alluxio/web/WebInterfaceGeneralServlet.java
@@ -15,7 +15,7 @@ import alluxio.Configuration;
 import alluxio.PropertyKey;
 import alluxio.RuntimeConstants;
 import alluxio.StorageTierAssoc;
-import alluxio.master.AlluxioMaster;
+import alluxio.master.AlluxioMasterService;
 import alluxio.master.file.FileSystemMaster;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.util.FormatUtils;
@@ -101,14 +101,14 @@ public final class WebInterfaceGeneralServlet extends HttpServlet {
 
   private static final long serialVersionUID = 2335205655766736309L;
 
-  private final transient AlluxioMaster mMaster;
+  private final transient AlluxioMasterService mMaster;
 
   /**
    * Creates a new instance of {@link WebInterfaceGeneralServlet}.
    *
    * @param master Alluxio master
    */
-  public WebInterfaceGeneralServlet(AlluxioMaster master) {
+  public WebInterfaceGeneralServlet(AlluxioMasterService master) {
     mMaster = master;
   }
 
@@ -167,7 +167,7 @@ public final class WebInterfaceGeneralServlet extends HttpServlet {
   private void populateValues(HttpServletRequest request) throws IOException {
     request.setAttribute("debug", Configuration.getBoolean(PropertyKey.DEBUG));
 
-    request.setAttribute("masterNodeAddress", mMaster.getMasterAddress().toString());
+    request.setAttribute("masterNodeAddress", mMaster.getRpcAddress().toString());
 
     request.setAttribute("uptime",
         WebUtils.convertMsToClockTime(System.currentTimeMillis() - mMaster.getStartTimeMs()));
