+++ tests/src/test/java/alluxio/master/lineage/LineageMasterClientRestApiTest.java
@@ -20,7 +20,7 @@ import alluxio.client.lineage.LineageContext;
 import alluxio.client.lineage.LineageFileSystem;
 import alluxio.job.CommandLineJob;
 import alluxio.job.JobConf;
-import alluxio.master.AlluxioMaster;
+import alluxio.master.AlluxioMasterService;
 import alluxio.rest.RestApiTest;
 import alluxio.rest.TestCase;
 import alluxio.wire.LineageInfo;
@@ -46,7 +46,7 @@ import javax.ws.rs.HttpMethod;
  * Test cases for {@link LineageMasterClientRestServiceHandler}.
  */
 public final class LineageMasterClientRestApiTest extends RestApiTest {
-  private AlluxioMaster mMaster;
+  private AlluxioMasterService mMaster;
   private LineageFileSystem mLineageClient;
 
   @Before
